pyproject.toml
[project]
name = "task-agent-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.18",
    "asyncio>=3.4.3",
    "fastapi>=0.100.0",
    "httpx>=0.24.1",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "uvicorn>=0.22.0",
]
